src/main/webapp/app/entities/corporate/update/corporate-update.component.ts
import { Component, OnInit, ElementRef } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder, Validators } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { ICorporate, Corporate } from '../corporate.model';
import { CorporateService } from '../service/corporate.service';
import { AlertError } from 'app/shared/alert/alert-error.model';
import { EventManager, EventWithContent } from 'app/core/util/event-manager.service';
import { DataUtils, FileLoadError } from 'app/core/util/data-util.service';
import { ICountry } from 'app/entities/country/country.model';
import { CountryService } from 'app/entities/country/service/country.service';

@Component({
  selector: 'jhi-corporate-update',
  templateUrl: './corporate-update.component.html',
})
export class CorporateUpdateComponent implements OnInit {
  isSaving = false;

  countriesSharedCollection: ICountry[] = [];

  editForm = this.fb.group({
    id: [],
    code: [null, [Validators.required, Validators.maxLength(5)]],
    name: [null, [Validators.required, Validators.maxLength(255)]],
    description: [null, [Validators.maxLength(500)]],
    logo: [null, [Validators.required]],
    logoContentType: [],
    contactPerson: [null, [Validators.maxLength(1000)]],
    employeeSize: [null, [Validators.required]],
    clientSize: [null, [Validators.required]],
    email: [
      null,
      [
        Validators.pattern(
          '^((&#34;[\\w-\\s]+&#34;)|([\\w-]+(?:\\.[\\w-]+)*)|(&#34;[\\w-\\s]+&#34;)([\\w-]+(?:\\.[\\w-]+)*))(@((?:[\\w-]+\\.)*\\w[\\w-]{0,66})\\.([a-z]{2,6}(?:\\.[a-z]{2})?)$)|(@\\[?((25[0-5]\\.|2[0-4][0-9]\\.|1[0-9]{2}\\.|[0-9]{1,2}\\.))((25[0-5]|2[0-4][0-9]|1[0-9]{2}|[0-9]{1,2})\\.){2}(25[0-5]|2[0-4][0-9]|1[0-9]{2}|[0-9]{1,2})\\]?$)'
        ),
      ],
    ],
    website: [
      null,
      [Validators.pattern('^(https?|ftp|file):\\\\/\\\\/[-a-zA-Z0-9+&amp;@#\\\\/%?=~_|!:,.;]*[-a-zA-Z0-9+&amp;@#\\\\/%=~_|]$')],
    ],
    country: [],
  });

  constructor(
    protected dataUtils: DataUtils,
    protected eventManager: EventManager,
    protected corporateService: CorporateService,
    protected countryService: CountryService,
    protected elementRef: ElementRef,
    protected activatedRoute: ActivatedRoute,
    protected fb: FormBuilder
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ corporate }) => {
      this.updateForm(corporate);

      this.loadRelationshipsOptions();
    });
  }

  byteSize(base64String: string): string {
    return this.dataUtils.byteSize(base64String);
  }

  openFile(base64String: string, contentType: string | null | undefined): void {
    this.dataUtils.openFile(base64String, contentType);
  }

  setFileData(event: Event, field: string, isImage: boolean): void {
    this.dataUtils.loadFileToForm(event, this.editForm, field, isImage).subscribe({
      error: (err: FileLoadError) =>
        this.eventManager.broadcast(
          new EventWithContent<AlertError>('balsamTestApp.error', { ...err, key: 'error.file.' + err.key })
        ),
    });
  }

  clearInputImage(field: string, fieldContentType: string, idInput: string): void {
    this.editForm.patchValue({
      [field]: null,
      [fieldContentType]: null,
    });
    if (idInput && this.elementRef.nativeElement.querySelector('#' + idInput)) {
      this.elementRef.nativeElement.querySelector('#' + idInput).value = null;
    }
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const corporate = this.createFromForm();
    if (corporate.id !== undefined) {
      this.subscribeToSaveResponse(this.corporateService.update(corporate));
    } else {
      this.subscribeToSaveResponse(this.corporateService.create(corporate));
    }
  }

  trackCountryById(index: number, item: ICountry): number {
    return item.id!;
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<ICorporate>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe(
      () => this.onSaveSuccess(),
      () => this.onSaveError()
    );
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(corporate: ICorporate): void {
    this.editForm.patchValue({
      id: corporate.id,
      code: corporate.code,
      name: corporate.name,
      description: corporate.description,
      logo: corporate.logo,
      logoContentType: corporate.logoContentType,
      contactPerson: corporate.contactPerson,
      employeeSize: corporate.employeeSize,
      clientSize: corporate.clientSize,
      email: corporate.email,
      website: corporate.website,
      country: corporate.country,
    });

    this.countriesSharedCollection = this.countryService.addCountryToCollectionIfMissing(this.countriesSharedCollection, corporate.country);
  }

  protected loadRelationshipsOptions(): void {
    this.countryService
      .query()
      .pipe(map((res: HttpResponse<ICountry[]>) => res.body ?? []))
      .pipe(
        map((countries: ICountry[]) => this.countryService.addCountryToCollectionIfMissing(countries, this.editForm.get('country')!.value))
      )
      .subscribe((countries: ICountry[]) => (this.countriesSharedCollection = countries));
  }

  protected createFromForm(): ICorporate {
    return {
      ...new Corporate(),
      id: this.editForm.get(['id'])!.value,
      code: this.editForm.get(['code'])!.value,
      name: this.editForm.get(['name'])!.value,
      description: this.editForm.get(['description'])!.value,
      logoContentType: this.editForm.get(['logoContentType'])!.value,
      logo: this.editForm.get(['logo'])!.value,
      contactPerson: this.editForm.get(['contactPerson'])!.value,
      employeeSize: this.editForm.get(['employeeSize'])!.value,
      clientSize: this.editForm.get(['clientSize'])!.value,
      email: this.editForm.get(['email'])!.value,
      website: this.editForm.get(['website'])!.value,
      country: this.editForm.get(['country'])!.value,
    };
  }
}
